refactor: adjust collection categories to be loaded from the main data source and remove from global
